requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.46.0
kiwisolver==1.4.5
matplotlib==3.8.2
MouseInfo==0.1.3
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
pytweening==1.0.7
PyYAML==6.0.1
six==1.16.0
